🧩 47KDMZ task: extract focused task-store services
