+++ .agentplane/tasks/202604172036-47KDMZ/pr/github-title.txt
@@ -1,0 +1,1 @@
+architecture/backend: Extract task-store services from shared monolith (47KDMZ)
