+++ .agentplane/tasks/202604172036-47KDMZ/pr/meta.json
@@ -1,0 +1,14 @@
+{
+  "base": "main",
+  "branch": "task/202604172036-47KDMZ/task-store-services",
+  "created_at": "2026-04-18T05:13:39.737Z",
+  "head_sha": "bdd6e9ee88aa87b648e0e28f22ed4ef2b8c7df92",
+  "last_verified_at": "2026-04-18T05:13:39.616Z",
+  "last_verified_sha": "bdd6e9ee88aa87b648e0e28f22ed4ef2b8c7df92",
+  "schema_version": 1,
+  "task_id": "202604172036-47KDMZ",
+  "updated_at": "2026-04-18T05:13:39.737Z",
+  "verify": {
+    "status": "pass"
+  }
+}
