+++ .agentplane/tasks/202604172036-47KDMZ/pr/review.md
@@ -1,0 +1,58 @@
+# PR Review
+
+Created: 2026-04-18T05:13:39.737Z
+Branch: task/202604172036-47KDMZ/task-store-services
+
+## Summary
+
+Extract task-store services from shared monolith
+
+Refactor commands/shared/task-store.ts into smaller read/write/projection services with explicit task backend capability boundaries so command flows stop depending on one large shared module.
+
+## Scope
+
+- In scope: Refactor commands/shared/task-store.ts into smaller read/write/projection services with explicit task backend capability boundaries so command flows stop depending on one large shared module.
+- Out of scope: unrelated refactors not required for "Extract task-store services from shared monolith".
+
+## Verification
+
+### Plan
+
+1. Run `bun run lint:core`. Expected: it succeeds and confirms the requested outcome for this task.
+2. Run `bun run test:fast`. Expected: it succeeds and confirms the requested outcome for this task.
+3. Review the changed artifact or behavior for the `code` task. Expected: the requested outcome is visible and matches the approved scope.
+4. Compare the final result against the task summary and touched scope. Expected: remaining follow-up is either resolved or explicit in ## Findings.
+
+### Current Status
+
+- State: ok
+- Note: Validated task-store service extraction: public task-store import surface stayed stable while implementation split into intent, README IO, store, and types modules; bun run lint:core and bun run test:fast both passed in the task worktree.
+
+## Risks
+
+- Risk level: not recorded
+- Breaking change: no
+
+### Rollback
+
+- Revert task-related commit(s).
+- Re-run required checks to confirm rollback safety.
+
+## Handoff Notes
+
+- No handoff notes recorded yet. Use `agentplane pr note ...` to append one.
+
+<!-- BEGIN AUTO SUMMARY -->
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-04-18T05:13:39.737Z
+- Branch: task/202604172036-47KDMZ/task-store-services
+- Head: bdd6e9ee88aa
+
+```text
+No changes detected.
+```
+
+</details>
+<!-- END AUTO SUMMARY -->
